refactor useAuth
