+++ desktop/src/hooks/index.ts
@@ -1,5 +1,6 @@
 export * from "./useAppCloseHandler";
+export * from './useAuth';
 export * from "./useFolderList/useFolderList";
-export * from "./useNavigateAway";
 export * from "./useNavigate/useNavigate";
+export * from "./useNavigateAway";
 export * from "./useReleaseNotes";
