+++ desktop/src/hooks/useAuth/AuthContext.tsx
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+export const AuthContext = createContext({
+  accessToken: undefined as string | undefined,
+  idToken: undefined as string | undefined,
+  isArchivist: false,
+});
