+++ desktop/src/hooks/useAuth/index.ts
@@ -1,0 +1,2 @@
+export * from './useAuth';
+export * from './AuthProvider';
