+++ desktop/src/hooks/useAuth/useAuth.tsx
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+import { AuthContext } from './AuthContext';
+
+
+export const useAuth = () => {
+	return useContext(AuthContext);
+};
